+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Add a `@cumulus/util/deprecate()` function to mark a piece of code as
     deprecated
 
+### Deprecated
+- **CUMULUS-965**
+  - Deprecated `@cumulus/common/concurrency.toPromise()` in favor of
+    [util.promisify()](https://nodejs.org/docs/v8.11.1/api/util.html#util_util_promisify_original)
+  - Deprecated `@cumulus/common/concurrency.promiseUrl()` in favor of
+    [got](https://www.npmjs.com/package/got)
+  - Deprecated `@cumulus/common/aws.syncUrl()`
+
 ## [v1.10.3] - 2018-10-31
 
 ### Added